// src/clean/presentation/pages/interfaces/landing-pages.interfaces.ts
export interface HookDescription {
  image: string;
  heading: string;
  content: string;
}

export interface HowItWorks {
  heading: string;
  content: string;
}

export interface UploadAbout {
  sectionTitle: string;
  image: string;
  content: string;
}

export interface FooterAd {
  backgroundimage: string;
  adTitle: string;
  adContent: string;
  callToAction: string;
}

export interface AdditionalContent {
  backgroundimage: string;
  heading: string;
  content: string;
}